Oilpan: have PendingScripts trace their script elements.
To complete the move to transition types for Element in html/ and dom/,
convert PendingScript and the script runner objects to trace their
embedded script element references. The script runner objects
(ScriptRunner and HTMLScriptRunner) are now garbage collected objects,
with PendingScript being kept in-line or as a part objects of those
runner objects.

To enable detaching of HTMLScriptRunner without having to rely on its
host (HTMLDocumentParser), the HTMLScriptRunner is instead made the
client of the resources it loads. This simplifies dependencies, but
requiring some follow-on changes to the HTMLScriptRunnerHost interface.

Besides script element references, convert a few other remaining Element
{Pass}RefPtr uses to transition types.

R=ager@chromium.org,haraken@chromium.org,eseidel@chromium.org
BUG=357163

Review URL: https://codereview.chromium.org/298863010

git-svn-id: svn://svn.chromium.org/blink/trunk@175127 bbb929c8-8fbe-4397-9dbb-9b2b20218538