Make the rebaseline log server resilient to large logs.
Right now when it exceeds the appengine storage limit for a single
entry it 500s. Instead, just create a new entry. This isn't totally
ideal. Ideally we'd create a new entry but link it to the old
one so we can show it correctly in the UI, but that's too much work.

NOTRY=true
BUG=378490

Review URL: https://codereview.chromium.org/304303003

git-svn-id: svn://svn.chromium.org/blink/trunk@175230 bbb929c8-8fbe-4397-9dbb-9b2b20218538