Stop firing orientationchange events at pages that are not visible
Stop firing 'orientationchange' events to pages that are not visible as per the
specification:
https://w3c.github.io/screen-orientation/#handling-screen-orientation-changes

This avoids having background pages handle 'orientationchange' events and
wasting battery.

As indicated in the specification, screen.orientation keeps returning the old
orientation if the orientation changes while the page is not visible.

R=mounir@chromium.org, jochen@chromium.org
BUG=162827
TEST=screen_orientation/page-visibility.html

Review URL: https://codereview.chromium.org/280283002

git-svn-id: svn://svn.chromium.org/blink/trunk@175264 bbb929c8-8fbe-4397-9dbb-9b2b20218538