Rename repaint to invalidation Part 1
Doing the rename from repaint to invalidation in one CL proved to be unwieldy.
This CL is the first step at the renaming. This moves a lot of pices that have
been added due to repaint-after-layout (although I'm leaving the flag named
as is to avoid Chromium side changes).

BUG=338691

Review URL: https://codereview.chromium.org/297903006

git-svn-id: svn://svn.chromium.org/blink/trunk@175291 bbb929c8-8fbe-4397-9dbb-9b2b20218538