DevTools: Event listener breakpoint should not stop if happened inside framework.
If we skip framework with StepInto after a stepping button was hit in front-end,
we should eventually stop somewhere in the user code. On the other hand, in case of
an Event breakpoint we should not stop at all, if the callback is handled inside the framework.

BUG=267592
R=yurys, pfeldman@chromium.org

Review URL: https://codereview.chromium.org/309013005

git-svn-id: svn://svn.chromium.org/blink/trunk@175471 bbb929c8-8fbe-4397-9dbb-9b2b20218538