Use WebViewClient instead of BlinkPlatform for screen lock.
Using BlinkPlatform hides the WebView that requested the lock, making it
very hard to have a clever handling of screen lock on the browser side
given that it's unknown whom and when requested the lock.

It will also allow locking the appropriate screen in a multi-screen
scenario.

BUG=162827

Review URL: https://codereview.chromium.org/315693002

git-svn-id: svn://svn.chromium.org/blink/trunk@175472 bbb929c8-8fbe-4397-9dbb-9b2b20218538