When we run sanitizers like ASAN and LSAN, we typically don't care about whether the tests run exactly like they do normally; we are just looking for errors that the sanitizers have detected.
This flag changes run-webkit-tests to ignore text and image failures and just look for crashes and timeouts (most crashes will be sanitizer-triggered).

In theory most of the time the tests *should* run the same way, but in practice we see a lot of noise that no one has the time to deal with, so this change should reduce a lot of the redness we see on the bots.

BUG=336971, 374043

Review URL: https://codereview.chromium.org/294973008

git-svn-id: svn://svn.chromium.org/blink/trunk@175488 bbb929c8-8fbe-4397-9dbb-9b2b20218538