Expose WebCrypto's algorithm normalization.
This exposes the WebCore::parseAlgorithm function through  blink::WebScriptBindings.

This will allow reusing the normalization in the implementation of the Chrome extension API enterprise.platformKeys.

To keep the exposure of types minimal, the dependency of parseAlgorithm on CryptoResult was removed.
On the way, renames parseAlgorithm to normalizeAlgorithm in order to match the WebCrypto spec and the filename.

BUG=364435

Review URL: https://codereview.chromium.org/295423004

git-svn-id: svn://svn.chromium.org/blink/trunk@175657 bbb929c8-8fbe-4397-9dbb-9b2b20218538