Emit object replacement char to IME
Currently Clank could only delete (with the backspace key) nodes that are
visible to the android IME, and since a replaced element, e.g. <img>, is not
visible to the IME, thus it could not be easily deleted using the backspace key.
Another side effect is that if two text nodes are separated by an image element
alone, then they appear as a single composition to the IME.

To fix these issues, this CL exposes a replaced element as a single object
replacement char 0xFFFC to InputMethodController. For more details please refer
to the atatched bug.

BUG=311448

Review URL: https://codereview.chromium.org/297003003

git-svn-id: svn://svn.chromium.org/blink/trunk@175819 bbb929c8-8fbe-4397-9dbb-9b2b20218538