Move server logging to rebaseline-o-matic.
It didn't really make sense to be part of auto-rebaseline except
for the special noneedsrebaseline entries, which I've decided are
not really useful anyways. More importantly, move it up means
that we actually log everything to the server, including the outer
git pull call and it means that we log asynchronously to the
auto-rebaseline command, which should make it a lot faster.

Also, make the git cl upload logging log stderr as well.

BUG=382248

Review URL: https://codereview.chromium.org/327443002

git-svn-id: svn://svn.chromium.org/blink/trunk@175828 bbb929c8-8fbe-4397-9dbb-9b2b20218538