Prepare SVGDocumentExtensions::m_elementDependencies for oilpan
|m_elementDependencies| had raw ptr to |SVGElement|s.
This was to avoid reference cycles, but we can convert this to strong references in oilpan.

If the whole document is dead, rebuild code is not called. This means that
rebuildElementReference is not called for referencing element, but the
referencing element must be in the same document and is going to die anyway.
If the whole document isn't dead (== the part of the document tree is going
away), the unregister/rebuild code is guaranteed to be called from removedFrom,
and rebuildElementReference is guaranteed to be triggered from there.

BUG=357163, 370834

Review URL: https://codereview.chromium.org/327473002

git-svn-id: svn://svn.chromium.org/blink/trunk@175880 bbb929c8-8fbe-4397-9dbb-9b2b20218538