Oilpan: have MediaController weakly refer to its media elements.
Simplify finalization of 'slaved' media elements by not requiring them
to unregister with their MediaController, but have it (the controller)
weakly refer to them instead.

Previously, a RefPtr<> was kept on the media element to make that
unregistration safe during finalization. This is no longer needed, and
a simpler Member reference from the media element is used instead.

R=haraken@chromium.org,ager@chromium.org,acolwell@chromium.org
BUG=357163

Review URL: https://codereview.chromium.org/314113008

git-svn-id: svn://svn.chromium.org/blink/trunk@175930 bbb929c8-8fbe-4397-9dbb-9b2b20218538