Reset computed size of the <canvas> FontDescription when setting CRC2D.font
Avoid inheriting the zoom-factor when setting the CanvasRenderingContext2D
'font' property to the same font as is computed for the <canvas> element.
In this case the font descriptions would be considered the same, and hence
the font size computed using the effective zoom of the page would be used.

BUG=381989

Review URL: https://codereview.chromium.org/326893003

git-svn-id: svn://svn.chromium.org/blink/trunk@175950 bbb929c8-8fbe-4397-9dbb-9b2b20218538