Clean up transform methods in GraphicsContext.
GraphicsContext::scale takes a FloatSize, which is not the way most
callers wish to invoke it. Here we change it to take x and y scale.

GraphicsContext::getCTM is not truly reflective of the tranforms
applied to get a pixel on the screen, as it ignores compositor applied
transforms. There are numerous locations where the CTM is used to test
pixel-level results, such as the width of a line. These tests are not
accurate in the presence of high DPI or other zoom factors. Each
location gets a FIXME indicating the problem.

R=fmalita,esprehn
BUG=383070,382491

Review URL: https://codereview.chromium.org/323013004

git-svn-id: svn://svn.chromium.org/blink/trunk@175960 bbb929c8-8fbe-4397-9dbb-9b2b20218538