SVG: no need to call |clearAnimatedType| at SVGAnimate{,Motion}Element dtor.
The deleted code was used to ensure the ASSERT(!isAnimating()) on ~SVGAnimatedProperty.
The assert checks that animation is always stopped when the element is dead. However, the assert is removed now, so we don't need to ensure it.

The deleted code |clearAnimatedType| in ~SVGAnimateElement was triggered when a document was unloaded and its <animate> element was destructed before its target svg element. This stopped animation on the target svg elements before the ASSERT run.

BUG=370834

Review URL: https://codereview.chromium.org/327463002

git-svn-id: svn://svn.chromium.org/blink/trunk@175989 bbb929c8-8fbe-4397-9dbb-9b2b20218538