Fix resolving protocol-agnostic <svg:image> URLs in about:blank iframes
When resolving a protocol-agnostic URL against a base of 'about:blank',
the result would be an invalid (and empty) URL, because of missing special-
casing of 'about:blank' - which is present in
Document::completeURLWithOverride.
Fix by modifying SVGImageLoader::sourceURI to always resolve the URL using
Document::completeURLWithOverride to not miss out on this special-casing
logic for 'about:blank' base URLs.

BUG=379100

Review URL: https://codereview.chromium.org/329943003

git-svn-id: svn://svn.chromium.org/blink/trunk@176043 bbb929c8-8fbe-4397-9dbb-9b2b20218538