Remove useless override of executionContext() in HTMLMediaElement
Remove useless override of executionContext() in HTMLMediaElement. This was
previously done to remove ambiguity because HTMLMediaElement inherited two
implementations of executionContext() from its base classes (Node and
ActiveDOMObject). This CL gets rid of the override and uses a "using"
statement instead to remove ambiguity, which is the proper way to do this.

The executionContext() override in Node can now be marked as FINAL as a
result.

R=tkent@chromium.org

Review URL: https://codereview.chromium.org/334063002

git-svn-id: svn://svn.chromium.org/blink/trunk@176167 bbb929c8-8fbe-4397-9dbb-9b2b20218538