[DirectWrite] Force fully aliased text for certain fonts at small sizes
Force the use of fully aliased text for certain common east asian fonts
at smaller font sizes to ensure that text is legible. This is only
required for fonts without an embedded bitmap at a the specific size and
the minimum size for enabling anti aliasing is set on a per font basis.

BUG=383090,383040
R=dglazkov@chromium.org, cpu@chromium.org

Review URL: https://codereview.chromium.org/339693002

git-svn-id: svn://svn.chromium.org/blink/trunk@176244 bbb929c8-8fbe-4397-9dbb-9b2b20218538